Restructure code base into different packages
