+++ src/main/java/luffy/command/AddDeadlineCommand.java
@@ -1,5 +1,13 @@
+package luffy.command;
+
 import java.io.IOException;
 import java.time.LocalDateTime;
+import luffy.exception.LuffyException;
+import luffy.task.TaskList;
+import luffy.task.Deadline;
+import luffy.ui.Ui;
+import luffy.storage.Storage;
+import luffy.parser.Parser;
 
 /**
  * Command to add a deadline task.
