+++ src/main/java/luffy/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package luffy.command;
+
+import luffy.task.TaskList;
+import luffy.ui.Ui;
+import luffy.storage.Storage;
+
 /**
  * Command to exit the application.
  */
