+++ src/main/java/luffy/exception/LuffyException.java
@@ -1,3 +1,5 @@
+package luffy.exception;
+
 public class LuffyException extends Exception {
     public LuffyException(String message) {
         super(message);
