+++ src/main/java/luffy/task/Task.java
@@ -1,3 +1,5 @@
+package luffy.task;
+
 public class Task {
     private String description;
     private boolean isDone;
